+++ modules/renter/contractor/contractmaintenance.go
@@ -606,6 +606,10 @@ func (c *Contractor) managedRenew(sc *proto.SafeContract, contractFunding types.
 // managedRenewContract will use the renew instructions to renew a contract,
 // returning the amount of money that was put into the contract for renewal.
 func (c *Contractor) managedRenewContract(renewInstructions fileContractRenewal, currentPeriod types.BlockHeight, allowance modules.Allowance, blockHeight, endHeight types.BlockHeight) (fundsSpent types.Currency, err error) {
+	if c.staticDeps.Disrupt("ContractRenewFail") {
+		err = errors.New("Renew failure due to dependency")
+		return
+	}
 	// Pull the variables out of the renewal.
 	id := renewInstructions.id
 	amount := renewInstructions.amount
@@ -1018,15 +1022,22 @@ func (c *Contractor) threadedContractMaintenance() {
 	if spending.TotalAllocated.Cmp(allowance.Funds) < 0 {
 		fundsRemaining = allowance.Funds.Sub(spending.TotalAllocated)
 	}
-	c.log.Debugln("Remaining funds in allowance:", fundsRemaining)
+	c.log.Debugln("Remaining funds in allowance:", fundsRemaining.HumanString())
 
-	// Register the AllowanceLowFunds alert if necessary.
+	// Register or unregister and alerts related to contract renewal or
+	// formation.
 	var registerLowFundsAlert bool
+	var renewErr error
 	defer func() {
 		if registerLowFundsAlert {
-			c.staticAlerter.RegisterAlert(modules.AlertIDRenterAllowanceLowFunds, AlertMSGAllowanceLowFunds, "", modules.SeverityWarning)
+			c.staticAlerter.RegisterAlert(modules.AlertIDRenterAllowanceLowFunds, AlertMSGAllowanceLowFunds, AlertCauseInsufficientAllowanceFunds, modules.SeverityWarning)
 		} else {
 			c.staticAlerter.UnregisterAlert(modules.AlertIDRenterAllowanceLowFunds)
+		}
+		if renewErr != nil {
+			c.staticAlerter.RegisterAlert(modules.AlertIDRenterContractRenewalError, AlertMSGFailedContractRenewal, renewErr.Error(), modules.SeverityError)
+		} else {
+			c.staticAlerter.UnregisterAlert(modules.AlertIDRenterContractRenewalError)
 		}
 	}()
 	// Go through the contracts we've assembled for renewal. Any contracts that
@@ -1056,6 +1067,7 @@ func (c *Contractor) threadedContractMaintenance() {
 		fundsSpent, err := c.managedRenewContract(renewal, currentPeriod, allowance, blockHeight, endHeight)
 		if err != nil {
 			c.log.Println("Error renewing a contract", renewal.id, err)
+			renewErr = errors.Compose(renewErr, err)
 		} else {
 			c.log.Println("Renewal completed without error")
 		}
@@ -1083,7 +1095,7 @@ func (c *Contractor) threadedContractMaintenance() {
 		// Skip this renewal if we don't have enough funds remaining.
 		c.log.Debugln("Attempting to perform a contract refresh:", renewal.id)
 		if renewal.amount.Cmp(fundsRemaining) > 0 || c.staticDeps.Disrupt("LowFundsRefresh") {
-			c.log.Println("skipping refresh because there are not enough funds remaining in the allowance", renewal.amount, fundsRemaining)
+			c.log.Println("skipping refresh because there are not enough funds remaining in the allowance", renewal.amount.HumanString(), fundsRemaining.HumanString())
 			registerLowFundsAlert = true
 			continue
 		}
@@ -1094,6 +1106,9 @@ func (c *Contractor) threadedContractMaintenance() {
 		fundsSpent, err := c.managedRenewContract(renewal, currentPeriod, allowance, blockHeight, endHeight)
 		if err != nil {
 			c.log.Println("Error refreshing a contract", renewal.id, err)
+			renewErr = errors.Compose(renewErr, err)
+		} else {
+			c.log.Println("Refresh completed without error")
 		}
 		fundsRemaining = fundsRemaining.Sub(fundsSpent)