+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Latest:
 **Key Updates**
  - Wallet can generate an address before it finishes scanning the blockchain
  - FUSE folders can now be mounted with 'AllowOther' as an option
+ - Added alerts for when contracts can't be renewed or refreshed
  - Smarter fund allocation when initially forming contracts
  - Decrease memory usage and cpu usage when uploading and downloading
  - When repairing files from disk, an integrity check is performed to ensure